🏷️ Refactor book settings types
