+++ server/types/book-settings.d.ts
@@ -1,0 +1,5 @@
+import type { Timestamp } from 'firebase-admin/firestore'
+
+export interface BookSettingsFirestoreData extends BaseBookSettingsData {
+  updatedAt?: Timestamp
+}
