+++ shared/types/book-settings.d.ts
@@ -1,0 +1,11 @@
+export interface BaseBookSettingsData {
+  'epub-cfi'?: string
+  'epub-fontSize'?: number
+  'epub-activeTTSElementIndex'?: number
+  'pdf-currentPage'?: number
+  'pdf-scale'?: number
+  'pdf-isDualPageMode'?: boolean
+  'pdf-isRightToLeft'?: boolean
+  'progress'?: number
+  'lastOpenedTime'?: number
+}
