+++ types/book-settings.d.ts
@@ -1,0 +1,14 @@
+export interface BookSettingsData extends BaseBookSettingsData {
+  updatedAt?: number
+}
+
+export type BookSettingKey =
+  | 'cfi'
+  | 'fontSize'
+  | 'activeTTSElementIndex'
+  | 'currentPage'
+  | 'scale'
+  | 'isDualPageMode'
+  | 'isRightToLeft'
+  | 'progress'
+  | 'lastOpenedTime'
